+++ test/token/http/grant/code.test.js
@@ -4,6 +4,7 @@ var $require = require('proxyquire');
 var expect = require('chai').expect;
 var sinon = require('sinon');
 var factory = require('../../../../com/token/http/grant/code');
+var oauth2orize = require('oauth2orize');
 
 
 describe('http/token/grant/code', function() {
@@ -148,6 +149,45 @@ describe('http/token/grant/code', function() {
         })
         .catch(done);
     }); // should not issue access token when authorization code not issued to client
+    
+    it('should not issue access token when redirect URI is not identical to initial authorization request', function(done) {
+      var codeSpy = sinon.stub();
+      var factory = $require('../../../../com/token/http/grant/code', {
+        'oauth2orize': { exchange: { code: codeSpy } }
+      });
+      
+      var acs = new Object();
+      acs.verify = sinon.stub().yieldsAsync(null, {
+        client: { id: 's6BhdRkqt3' },
+        redirectURI: 'https://client.example.org/cb',
+        user: { id: '248289761001' },
+        scope: [ 'profile', 'email' ]
+      });
+      var ats = new Object();
+      ats.issue = sinon.stub().yieldsAsync(null, '2YotnFZFEjr1zCsicMWpAA');
+      
+      factory(ats, acs, logger, container)
+        .then(function(exchange) {
+          
+          var client = {
+            id: 's6BhdRkqt3',
+            name: 'Example Client',
+            redirectURIs: [ 'https://client.example.org/cb' ]
+          };
+          
+          var issue = codeSpy.getCall(0).args[0];
+          issue(client, 'SplxlOBeZQQYbYS6WxSbIA', 'https://client.example.org/callback', {}, {}, function(err, token) {
+            expect(err).to.be.an.instanceOf(oauth2orize.TokenError);
+            expect(err.message).to.equal('Mismatched redirect URI');
+        
+            expect(acs.verify).to.be.calledOnce;
+            expect(acs.verify.getCall(0).args[0]).to.equal('SplxlOBeZQQYbYS6WxSbIA');
+            expect(ats.issue).to.not.be.called;
+            done();
+          });
+        })
+        .catch(done);
+    }); // should not issue access token when redirect URI is not identical to initial authorization request
     
   }); // creating exchange